fix(app): 修复 core/app 模块全部 Kotlin 编译错误
并行修复 25 个文件的编译错误, 按模块分组:

[Audio/Video 预览]
- AudioPreviewFragment.kt: Float.toShort() 废弃, 改为 toInt().toShort()
- AudioVisualizer.kt: Visualizer.setDataListener 不存在, 改为 setDataCaptureListener
- VideoGestureHandler.kt: by remember 缺 mutableStateOf 委托
- VideoPlaybackController.kt: onTracksChanged 签名缺 tracks: Tracks 参数
- VideoPreviewFragment.kt: WindowManager.LayoutParams.brightness 改为 screenBrightness

[Web 预览]
- WebPreviewFragment.kt: WebPreviewScreen 顶层 composable 无法访问
  Fragment 的 backendSession 字段, 改用局部 currentSession state

[Debugger connection]
- ConnectionBackedDebugger.kt: catch (Throwable t) → catch (t: Throwable)
- ConnectionRetryPolicy.kt: Result.Failure<T> 非泛型, 改为 Result<T>
- AidlHostLauncher.kt: EXTRA_PORT 在 companion, 改为限定引用
- DebugTarget.kt: mainActivity 改为可空 (AppReadyAutoConnect 传 null)
- BreakpointInjectionCodeGen.kt: activityClass→mainActivity, $转义
- HostBridgeServer.kt: LocalServerSocket 无 isClosed, parseHello 参数可空
- AdbForwardConnection.kt: resolve() 返回类型, smart cast 修复
- InnetVmSocksConnection.kt: 同 resolve() 返回类型
- RootConnection.kt: suBinary→suBinPath (RootConfig 字段名)
- ShizukuConnection.kt: lambda 显式类型, runBlocking 桥接 suspend,
  server.receive()→accept()
- ProcessRunner.kt: arrayOfNulls 传 null 给非空 Int, 改为 1
- HostPluginProbe.kt: @Volatile 不能用于局部变量, ServiceConnection
  非 SAM 改 object 表达式, processName→processNameSuffix,
  unbindUserService 补 builder 参数
- ShizukuBinderClient.kt: 同 HostPluginProbe, checkPermission→
  checkRemotePermission
- ShizukuFdTransporter.kt: getOption/setOption 签名可空标注错误
- ShizukuSubPathCapabilities.kt: DebugTarget 无 adbSerial, 改为 _ -> true
- Socks5Client.kt: ILogger.ROOT 实际类型是 org.slf4j.Logger

[Editor + Preferences]
- EditorHandlerActivity.kt: needsInjector→entry.category==BROWSER,
  showForEdit→showForExisting, showSidebar→show
- debuggerPrefExts.kt: `is`= 参数名不存在, 改为 _isChecked=

+++ core/app/src/main/java/com/itsaky/androidide/fragments/editor/audio/AudioPreviewFragment.kt
@@ -658,7 +658,7 @@ private fun EqualizerBandSlider(
         ) {
             FrostedSlider(
                 value = currentValue,
-                onValueChange = { onLevelChange(it.toShort()) },
+                onValueChange = { onLevelChange(it.toInt().toShort()) },
                 valueRange = valueRange,
                 enabled = enabled,
                 modifier = Modifier
